+++ functional/order-service/src/main/java/com/microservice/order/dto/request/GetOrderDetailsRequest.java
@@ -2,6 +2,7 @@ package com.microservice.order.dto.request;
 
 
 import com.microservice.order.dto.InvoiceDTO;
+import lombok.Data;
 
 import java.math.BigDecimal;
 import java.util.Date;
@@ -10,15 +11,7 @@ import java.util.List;
  * Created by dendy-prtha on 01/07/2019.
  * get order details  DTO
  */
-
+@Data
 public class GetOrderDetailsRequest {
     private int orderId;
-
-    public int getOrderId() {
-        return orderId;
-    }
-
-    public void setOrderId(int orderId) {
-        this.orderId = orderId;
-    }
 }
