+++ functional/order-service/src/main/java/com/microservice/order/dto/response/AddOrderItemToOrderResponse.java
@@ -1,29 +1,15 @@
 package com.microservice.order.dto.response;
 
+import lombok.Data;
+
 /*
  * Created by dendy-prtha on 17/05/2019.
  *
  */
-
+@Data
 public class AddOrderItemToOrderResponse {
 
     private String status;
 
     private String itemId;
-
-    public String getStatus() {
-        return status;
-    }
-
-    public void setStatus(String status) {
-        this.status = status;
-    }
-
-    public String getItemId() {
-        return itemId;
-    }
-
-    public void setItemId(String itemId) {
-        this.itemId = itemId;
-    }
 }
