--- functional/order-service/src/main/java/com/microservice/order/repository/dao/Dao.java
@@ -1,48 +1,0 @@
-package com.microservice.order.repository.dao;
-
-import java.util.List;
-import java.util.Optional;
-
-/**
- * Created by dendy-prtha on 01/03/2019.
- * This interface provides generic CRUD operation. See
- * <a href="https://en.wikipedia.org/wiki/Create,_read,_update_and_delete>CRUD</a>
- * Its uses generic to provide type-safe operation.
- */
-
-public interface Dao<T, ID>
-{
-    /**
-     * Find single record by id.
-     * @param id id
-     * @return record wrapped by {@link Optional}
-     */
-    Optional<T> findById(ID id);
-
-    /**
-     * Find all records
-     * @return list of records
-     */
-    List<T> findAll();
-
-    /**
-     * Persist record to database
-     * @param o record
-     * @return true if operation success
-     */
-    boolean save(T o);
-
-    /**
-     * Persist existing record to database
-     * @param o record
-     * @return true if operation success
-     */
-    boolean update(T o);
-
-    /**
-     * Remove existing record from database
-     * @param o record
-     * @return true if operaion success
-     */
-    boolean delete(T o);
-}
