--- functional/order-service/src/main/java/com/microservice/order/repository/dao/PaymentMethodDao.java
@@ -1,11 +1,0 @@
-package com.microservice.order.repository.dao;
-
-import com.microservice.order.model.PaymentMethod;
-
-/*
- * Created by dendy-prtha on 11/06/2019.
- * Video Dao
- */
-
-public interface PaymentMethodDao extends Dao<PaymentMethod, Integer> {
-}
